Created XML based configuration loading
Added interface for instance-based configuration for robot components
which may need tweaking that doesn't require a full rebuild.  It
features saving and loading, so smart dashboard data points can be saved
back to the file for easy PID tuning.  This is also designed to be used
hierarchially, but does not enforce it.